Module fits_rs::types

The types modules describes all the structures to express FITS files.

Structs

BlankRecord
A unit tuple that will act as a placeholder for blank records.

Fits
Representation of a FITS file.

KeywordRecord
A keyword record contains information about a FITS header.

PrimaryHeader
The primary header of a FITS file.

Enums

Keyword
The various keywords that can be found in headers.

ParseKeywordError
Problems that could occur when parsing a str are enumerated here.